fix swayidle build by disabling bash completions
Otherwise, this is failing with:
`Read-only file system: '/usr/share/bash-completion/completions/swayidle'`
both on my system and on CI.
The issue seems that if [bash-completion](https://github.com/scop/bash-completion)
is available, it uses that to get the dir for bash completions, which appears to
ignore the prefix and tells to install to `/usr/share`.
Without `bash-completion`, it falls back to datadir, which correctly uses `/app/share`.

Also, we don't need bash completions.